requirements.txt
aiohttp
async-timeout
attrs
beautifulsoup4
certifi
chardet
discord
discord.py # pip install -U "discord.py[voice]"
idna
# libxml2
# libxslt
# lxml~=4.6.2
multidict
numpy
pandas
python-dateutil
python-dotenv
pytz
requests
six
soupsieve
tabulate
TenGiphPy
# tenorscrap~=0.0.3 # pip install git+https://github.com/suarasiy/tenorscrap
urllib3
yarl
youtube-data-api
youtube-dl
youtube-search
youtube-search-python
yt_dlp
feedparser
basc_py4chan
pillow
eyed3
tinydb
imagetext_py
duckduckgo_search
